requirements.txt
anyio==4.9.0
astor==0.8.1
attrdict==2.0.1
babel==2.17.0
bce-python-sdk==0.9.35
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.1.1
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fire==0.7.0
Flask==3.1.1
flask-babel==4.0.0
flatbuffers==25.2.10
fonttools==4.58.2
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.6.2
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.9.4
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.5
onnx==1.17.0
onnxruntime==1.19.2
onnxslim==0.1.57
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==2.7.0.0
paddlepaddle==2.5.2
pandas==2.3.0
pdf2docx==0.5.8
pillow==11.2.1
premailer==3.10.0
protobuf==3.20.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
PyMuPDF==1.20.2
pyparsing==3.2.3
pyreadline3==3.5.4
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
rarfile==4.2
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.7
shiboken6==6.5.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sympy==1.14.0
termcolor==3.1.0
thop==0.1.1.post2209072238
tifffile==2024.8.30
torch==1.13.1
torchvision==0.14.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.151
ultralytics-thop==2.0.14
urllib3==2.4.0
visualdl==2.5.3
Werkzeug==3.1.3
zipp==3.22.0
